+++ HelloJava/src/com/paran/chapter5/ControlStateMentDemo.java
@@ -58,33 +58,4 @@ public class ControlStateMentDemo {
 //private : 내 클래스 안에서만
 //public: 전 월드에서 사용 가능. 
 
-	class Student
-	{
-		private String name;
-		private String StudentId;
-		private boolean isMilSer; //전역여부
-		
-//		public Student(){
-//		}
-		
-		public Student(String name,String StudentId,boolean isMilSer){
-			this.name = name;
-			this.StudentId = StudentId;
-			this.isMilSer = isMilSer;
-		}
-		public void showInfo(){
-		System.out.println("이름:" + name+ "학번 " + StudentId);
-		}
-		
-		//getter function  Statement 
-		
-		
-		
-		public boolean getIsMilser(){
-			return isMilSer;
-		}
-		
-		public void setIsMilser(boolean isMilSer) {
-			
-		}
-	}
+
