+++ HelloJava/src/com/paran/chapter7/ClassInstanceDemo.java
@@ -1,0 +1,13 @@
+package com.paran.chapter7;
+import com.paran.chapter5.Student;
+
+
+public class ClassInstanceDemo {
+	public static void main(String[] args) {
+		Student obj = null;
+		if (obj != null) System.out.println(obj.getIsMilser());
+		
+		//지역변수는 초기화를 선언 필요하다.!! 
+	}
+
+}
